Move Types.ts from autotest package to the common package and update references
