+++ packages/portal/backend/src/Types.ts
@@ -1,4 +1,4 @@
-import {IAutoTestResult} from "../../../autotest/src/Types";
+import {IAutoTestResult} from "../../../common/types/AutoTestTypes";
 
 /**
  * These types are the storage-specific types used by the backend.
