+++ packages/portal/backend/src/controllers/ResultsController.ts
@@ -1,7 +1,7 @@
 import Log from "../../../../common/Log";
 
 import {DatabaseController} from "./DatabaseController";
-import {IAutoTestResult} from "../../../../autotest/src/Types";
+import {IAutoTestResult} from "../../../../common/types/AutoTestTypes";
 import {RepositoryController} from "./RepositoryController";
 import {Result} from "../Types";
